Implement project structure to organize modules, refactor some components
