+++ app/components/YourTitle/index.js
@@ -1,0 +1,3 @@
+import YourTitle from "./YourTitle";
+
+export default YourTitle;
