+++ app/package.json
@@ -1,0 +1,3 @@
+{
+ "name": "app"
+}
