+++ app/screens/Home/index.js
@@ -1,0 +1,3 @@
+import HomeScreen from './HomeScreen'
+
+export default HomeScreen
