+++ app/screens/LogList/index.js
@@ -1,0 +1,3 @@
+import LogList from "./LogList";
+
+export default LogList;
